ai_chat/requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
anthropic>=0.8.1
openai>=1.9.0
duckduckgo-search>=4.2
brave-search==0.1.8
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
serpapi>=0.1.5
sqlparse>=0.4.4
tabulate>=0.9.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio>=3.4.3
beautifulsoup4>=4.12.0
markdown>=3.5.0
mistralai>=0.0.12
huggingface-hub>=0.20.0
tiktoken>=0.5.0
groq>=0.4.0
pygame>=2.6.1
swarm @ git+https://github.com/openai/swarm.git